添加生活馆复制微信，品牌馆选择优惠券的bug，提交订单的高度bug，大B展示了开馆指引的bug解决C

+++ lexi/Feature/LivingHall/View/THNLivingHallRecommendTableViewCell.m
@@ -104,7 +104,9 @@ static NSString *const kUrlProductUserLike = @"/product/userlike";
 }
 
 - (void)deleteProduct {
-    self.deleteProductBlock(self);
+    if (self.deleteProductBlock) {
+         self.deleteProductBlock(self);
+    }
 }
 
 
@@ -124,7 +126,9 @@ static NSString *const kUrlProductUserLike = @"/product/userlike";
 }
 
 - (IBAction)shareButton:(id)sender {
-    
+    if (self.shareProductBlock) {
+        self.shareProductBlock();
+    }
 }
 
 - (void)setProductModel:(THNProductModel *)productModel {
